+++ lib/app_base.dart
@@ -7,5 +7,6 @@ export 'src/analytics.dart';
 export 'src/components/change_based_text_field.dart';
 export 'src/components/padded_card.dart';
 export 'src/di.dart';
+export 'src/future_widget/firestore_query_view.dart';
 export 'src/stream_widget/firestore_query_view.dart';
 export 'src/stream_widget/stream_widget.dart';